pyproject.toml
[project]
name = "bank-automation-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Hashir",email = "hashirventhodi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (==0.103.1)",
    "uvicorn (==0.23.2)",
    "pydantic (==2.3.0)",
    "python-dotenv (==1.0.0)",
    "sqlalchemy (==2.0.20)",
    "alembic (==1.12.0)",
    "psycopg2-binary (==2.9.7)",
    "pdfminer-six (==20221105)",
    "pypdf2 (==3.0.1)",
    "tabula-py (==2.7.0)",
    "pdf2image (==1.16.3)",
    "pytesseract (==0.3.10)",
    "opencv-python (==4.11.0.86)",
    "google-cloud-vision (==3.4.3)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "openpyxl (==3.1.2)",
    "regex (==2023.8.8)",
    "requests (==2.31.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "fastapi-jwt-auth (==0.5.0)",
    "tqdm (==4.66.1)",
    "loguru (==0.7.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.2"
pytest-cov = "4.1.0"

